requirements.txt
pandas>=2.0
numpy>=1.26
python-dotenv>=1.0